Added fish_emitter from TR3
- Now fish take the room ambient as texture color.
- Now fish behaviour can be changed by OCB.

+++ TOMB4-custom/TOMB4/game/sound.cpp
@@ -680,7 +680,7 @@ void SoundSystem::PlaySoundTrack(std::string track, SOUND_TRACK_TYPE mode, QWORD
 		// Only activates if no custom position is passed as argument.
 		if (!position)
 		{
-			QWORD newPos = BASS_ChannelGetLength(stream, BASS_POS_BYTE) * (static_cast<float>(GetRandomControl()) / static_cast<float>(RAND_MAX));
+			QWORD newPos = BASS_ChannelGetLength(stream, BASS_POS_BYTE) * QWORD(static_cast<float>(GetRandomControl()) / static_cast<float>(RAND_MAX));
 			BASS_ChannelSetPosition(stream, newPos, BASS_POS_BYTE);
 		}
 	}
